Improve computation of operation hashes.
