+++ src/fields/field_fe.hh
@@ -46,6 +46,37 @@
 
 
 
+/**
+ * Helper class that holds data used in FieldFE operations
+ * (see source file src/fields/field_fee.hh).
+ */
+class FieldFeOpData {
+public:
+    /// Constructor
+    FieldFeOpData(std::shared_ptr<DOFHandlerMultiDim> dh, VectorMPI data_vec)
+    : dh_(dh), data_vec_(data_vec) {}
+
+    inline std::shared_ptr<DOFHandlerMultiDim> dh() const {
+    	return dh_;
+    }
+
+    inline VectorMPI data_vec() const {
+    	return data_vec_;
+    }
+
+    bool operator==(const FieldFeOpData &other)
+    {
+        return (dh_->hash() == other.dh_->hash()) &&
+               (data_vec_.size() == other.data_vec_.size());
+    }
+
+private:
+    std::shared_ptr<DOFHandlerMultiDim> dh_;
+    VectorMPI data_vec_;
+};
+
+
+
 namespace internal {
 
 template<unsigned int dim>
@@ -499,6 +530,23 @@ private:
     /// Registrar of class to factory
     static const int registrar;
 };
+
+
+namespace std {
+
+/// Template specialization of std::hash for FieldFeOpData
+template<>
+struct hash< FieldFeOpData > {
+    std::size_t operator()(const FieldFeOpData &op_data) const noexcept {
+        std::size_t h1 = std::hash<std::size_t>{}( op_data.dh()->hash() );
+        std::size_t h2 = std::hash<uint>{}( op_data.data_vec().size() );
+
+        // hash combine
+        return h1 ^ (h2 + 0x9e3779b9 + (h1 << 6) + (h1 >> 2));
+    }
+};
+
+} // namespace std
 
 
 /** Create FieldFE from dof handler */
